requirements.txt
flask
gunicorn
python-shogi